+++ packages/accounts-controller/src/AccountsController.test.ts
@@ -373,7 +373,30 @@ describe('AccountsController', () => {
     afterEach(() => {
       jest.clearAllMocks();
     });
-    it('should only update if the keyring is unlocked', async () => {
+    it('should not update state when only keyring is unlocked without any keyrings', async () => {
+      const messenger = buildMessenger();
+      const accountsController = setupAccountsController({
+        initialState: {
+          internalAccounts: {
+            accounts: {},
+            selectedAccount: '',
+          },
+        },
+        messenger,
+      });
+
+      messenger.publish(
+        'KeyringController:stateChange',
+        { isUnlocked: true, keyrings: [] },
+        [],
+      );
+
+      const accounts = accountsController.listAccounts();
+
+      expect(accounts).toStrictEqual([]);
+    });
+
+    it('should only update if the keyring is unlocked and when there are keyrings', async () => {
       const messenger = buildMessenger();
 
       const mockNewKeyringState = {
@@ -1850,6 +1873,7 @@ describe('AccountsController', () => {
       jest.spyOn(AccountsController.prototype, 'updateAccounts');
       jest.spyOn(AccountsController.prototype, 'getAccountByAddress');
       jest.spyOn(AccountsController.prototype, 'getSelectedAccount');
+      jest.spyOn(AccountsController.prototype, 'getAccount');
     });
 
     describe('setSelectedAccount', () => {
@@ -1987,6 +2011,31 @@ describe('AccountsController', () => {
 
         const account = messenger.call('AccountsController:getSelectedAccount');
         expect(accountsController.getSelectedAccount).toHaveBeenCalledWith();
+        expect(account).toStrictEqual(mockAccount);
+      });
+    });
+
+    describe('getAccount', () => {
+      it('should get account by id', async () => {
+        const messenger = buildMessenger();
+
+        const accountsController = setupAccountsController({
+          initialState: {
+            internalAccounts: {
+              accounts: { [mockAccount.id]: mockAccount },
+              selectedAccount: mockAccount.id,
+            },
+          },
+          messenger,
+        });
+
+        const account = messenger.call(
+          'AccountsController:getAccount',
+          mockAccount.id,
+        );
+        expect(accountsController.getAccount).toHaveBeenCalledWith(
+          mockAccount.id,
+        );
         expect(account).toStrictEqual(mockAccount);
       });
     });